feat: Add read function to comment module
This commit adds a new function called "read" to the comment module in the MongoDB module. The read function retrieves a comment from the database based on its ID. If the comment is not found, an error is thrown. This feature will enhance the user's ability to retrieve specific comments from the database.