feat: Add confirmation ID field to joinRequest schema
This change adds a new field, `confirmId`, to the `joinRequest` schema
in order to store a unique confirmation ID for each join request. The
confirmation ID is generated using the `uuid` library. This addition
enhances the functionality of the joinRequest feature by providing a
unique identifier for confirming join requests. The confirmation ID can
be used for validating and tracking the status of a join request.